♻️ chore: Split code by files from BalanceProvider
